A variety of source code tweaks to try and get things working with newer boost (using 1.47 at the moment), glog, gflags, etc...  I can't quite get things working yet, so these may or may not be fully correct...

+++ crypto/evp.hpp
@@ -55,7 +55,7 @@ class EVP {
   template <class Type>
   inline Type digest() const {
     Type ret;
-    for (int i = 0; i < min(md_len_, sizeof(Type)); ++i) {
+    for (int i = 0; i < min(md_len_, (unsigned int)sizeof(Type)); ++i) {
       ret = (ret) << 8 | md_value_[i];
     }
     return ret;
